Added scripts and objects
Finished a section on the debug script to allow player to return to a specific point on the map.

Need to fix so the debug button can be used while the player is jumping.

Added a weapon standin for the player and getting ready to program it.

Added an area for the player to prctice the jump mechanic.

Added a sprint to the player to go faster when key is held down and return to base variable when the key is released.